Move board adapter tests out of main module
